refactor(recipes-data-access): group services and interceptors into separate folders
