+++ libs/recipes/data-access-recipes/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/recipes-data-access-recipes.module';
 export * from './lib/models';
 export * from './lib/store';
-export * from './lib/http-api.interceptor';
-export * from './lib/http-error.interceptor';
-export { RecipeHttpService } from './lib/recipe-http.service';
+export * from './lib/services';
+export * from './lib/interceptors';
