+++ libs/recipes/data-access-recipes/src/lib/interceptors/http-api.interceptor.spec.ts
@@ -3,11 +3,11 @@ import { TestBed } from '@angular/core/testing';
 import { HttpApiInterceptor } from './http-api.interceptor';
 
 describe('HttpApiInterceptor', () => {
-  beforeEach(() => TestBed.configureTestingModule({
-    providers: [
-      HttpApiInterceptor
-      ]
-  }));
+  beforeEach(() =>
+    TestBed.configureTestingModule({
+      providers: [HttpApiInterceptor],
+    })
+  );
 
   it('should be created', () => {
     const interceptor: HttpApiInterceptor = TestBed.inject(HttpApiInterceptor);
