+++ libs/recipes/data-access-recipes/src/lib/interceptors/http-api.interceptor.ts
