+++ libs/recipes/data-access-recipes/src/lib/interceptors/http-error.interceptor.spec.ts
@@ -3,14 +3,15 @@ import { TestBed } from '@angular/core/testing';
 import { HttpErrorInterceptor } from './http-error.interceptor';
 
 describe('HttpErrorInterceptor', () => {
-  beforeEach(() => TestBed.configureTestingModule({
-    providers: [
-      HttpErrorInterceptor
-      ]
-  }));
+  beforeEach(() =>
+    TestBed.configureTestingModule({
+      providers: [HttpErrorInterceptor],
+    })
+  );
 
   it('should be created', () => {
-    const interceptor: HttpErrorInterceptor = TestBed.inject(HttpErrorInterceptor);
+    const interceptor: HttpErrorInterceptor =
+      TestBed.inject(HttpErrorInterceptor);
     expect(interceptor).toBeTruthy();
   });
 });
