+++ libs/recipes/data-access-recipes/src/lib/interceptors/index.ts
@@ -1,0 +1,2 @@
+export { HttpApiInterceptor } from './http-api.interceptor';
+export { HttpErrorInterceptor } from './http-error.interceptor';
