+++ libs/recipes/data-access-recipes/src/lib/services/index.ts
@@ -1,0 +1,1 @@
+export { RecipeHttpService } from './recipe-http.service';
